+++ samples/boards/nrf52/mesh/onoff_level_lighting_vnd_app/CMakeLists.txt
@@ -10,6 +10,7 @@ target_sources(app PRIVATE
 	       src/mesh/ble_mesh.c
 	       src/mesh/device_composition.c
 	       src/mesh/publisher.c
+	       src/mesh/state_binding.c
 	       )
 
 zephyr_include_directories(
