+++ samples/boards/nrf52/mesh/onoff_level_lighting_vnd_app/src/main.c
@@ -12,6 +12,7 @@
 #include "ble_mesh.h"
 #include "device_composition.h"
 #include "publisher.h"
+#include "state_binding.h"
 
 struct device *led_device[4];
 struct device *button_device[4];
@@ -89,6 +90,55 @@ static void gpio_init(void)
 	gpio_init_callback(&button_cb[3], button_pressed, BIT(SW3_GPIO_PIN));
 	gpio_add_callback(button_device[3], &button_cb[3]);
 	gpio_pin_enable_callback(button_device[3], SW3_GPIO_PIN);
+}
+
+void light_default_status_init(void)
+{
+	/* Assume vaules are retrived from Persistence Storage (Start).
+	 * These had saved by respective Setup Servers.
+	 */
+	gen_power_onoff_srv_user_data.onpowerup = STATE_DEFAULT;
+
+	light_lightness_srv_user_data.light_range_min = LIGHTNESS_MIN;
+	light_lightness_srv_user_data.light_range_max = LIGHTNESS_MAX;
+	light_lightness_srv_user_data.def = LIGHTNESS_MAX;
+
+	/* Following 2 values are as per specification */
+	light_ctl_srv_user_data.temp_range_min = TEMP_MIN;
+	light_ctl_srv_user_data.temp_range_max = TEMP_MAX;
+
+	light_ctl_srv_user_data.temp_def = TEMP_MIN;
+	/* (End) */
+
+	/* Assume following values are retrived from Persistence
+	 * Storage (Start).
+	 * These values had saved before power down.
+	 */
+	light_lightness_srv_user_data.last = LIGHTNESS_MAX;
+	light_ctl_srv_user_data.temp_last = TEMP_MIN;
+	/* (End) */
+
+	light_ctl_srv_user_data.temp = light_ctl_srv_user_data.temp_def;
+
+	if (gen_power_onoff_srv_user_data.onpowerup == STATE_OFF) {
+		gen_onoff_srv_root_user_data.onoff = STATE_OFF;
+		state_binding(ONOFF, ONOFF_TEMP);
+	} else if (gen_power_onoff_srv_user_data.onpowerup == STATE_DEFAULT) {
+		gen_onoff_srv_root_user_data.onoff = STATE_ON;
+		state_binding(ONOFF, ONOFF_TEMP);
+	} else if (gen_power_onoff_srv_user_data.onpowerup == STATE_RESTORE) {
+		/* Assume following values is retrived from Persistence
+		 * Storage (Start).
+		 * This value had saved before power down.
+		 */
+		gen_onoff_srv_root_user_data.onoff = STATE_ON;
+		/* (End) */
+
+		light_ctl_srv_user_data.temp =
+			light_ctl_srv_user_data.temp_last;
+
+		state_binding(ONPOWERUP, ONOFF_TEMP);
+	}
 }
 
 void update_light_state(void)
