+++ samples/boards/nrf52/mesh/onoff_level_lighting_vnd_app/src/mesh/device_composition.h
@@ -73,8 +73,8 @@ struct light_lightness_state {
 	u16_t def;
 
 	u8_t status_code;
-	u16_t lightness_range_min;
-	u16_t lightness_range_max;
+	u16_t light_range_min;
+	u16_t light_range_max;
 
 	u8_t last_tid;
 	u16_t last_tx_addr;
@@ -100,7 +100,10 @@ struct light_ctl_state {
 };
 
 extern struct generic_onoff_state gen_onoff_srv_root_user_data;
+extern struct generic_level_state gen_level_srv_root_user_data;
+extern struct generic_onpowerup_state gen_power_onoff_srv_user_data;
 extern struct light_lightness_state light_lightness_srv_user_data;
+extern struct light_ctl_state light_ctl_srv_user_data;
 extern struct generic_level_state gen_level_srv_s0_user_data;
 
 extern struct bt_mesh_model root_models[];
